make sure that the client wont send data to converter in case of error
HTTP client implementation often perform:

connect()
send()
read()

The hook to check for the convert header was previously on read() this
means that data was still sent to the converter even if it sent the Error TLV.

This change adds a check at the end of connect() for blocking sockets (wget)
as well as in send() for non-blocking sockets (curl).
Note that I tried to hook poll() without success. Returning POLLERR was ignored
by curl.

The code introduces a possible peek() instead of read of the Convert HDR+TLV.
It is not yet used but could be useful for the future.

Change-Id: If11ca8f13434bca9c228633bcb0f4f40f9f01cfd
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Gregory Vander Schueren <gregory.vanderschueren@tessares.net>