convert_util: expose CONVERT_ALIGN() macro
This macro can be useful for client code to allocate space for
Convert messages as the space should indeed account for the 4-bytes
alignment. This would be the case for example when calling
convert_write(): the function would return an error if the provided
buffer is not large enough for the 4-bytes alignment.

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>