convert_client: non-matching IP versions
This patch adds support in convert_client for non-matching IP versions
on upstream and downstream TCP connections.

Up to now, the convert_client was making the incorrect assumption that
the client would only reach IPv4 hosts when using a Converter addressable
over IPv4 (and similarly for IPv6). However, it should be possible to
reach both an IPv4 and IPv6 end server through a Converter addressable
over IPv4 (same goes for IPv6). IP versions used on upstream and
downstream don't have to match.

In particular, this was posing serious problems when using the toy
convert_client with an end-server such as "google.com" which resolves
to both IPv4 and IPv6. While the convert_client only accepts a single
converter address via the CONVERT_ADDR env variable, curl/wget will be
doing happy eyeball and will thus necessarily open a socket with an
IP family different from the converter address, hence crashing the
client.

Signed-off-by: Gregory Vander Schueren <gregory.vanderschueren@tessares.net>