headers: include TLV base struct in other ones
All Convert TLVs have the same base including the type and the length.

In the code, we have the base TLV structure that is extended by others.
We cast the same pointer from the base TLV to extended versions. It is
OK to do that because each TLV has the same two variables of the same
length as base.

Better to explicitly mark it like that in the code then.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>